Fixes for cmv4 (Brackets Sprint 38):
- Change to Alt+Drag to avoid interfering with Ctrl+Drag to add another
selection range. (If we cancel the ctrl+mousedown, it's impossible to even
drop a second cursor on a number; if we don't cancel the mousedown and only
cancel mousemoves, CM starts trying to drag out a selection and there's
flicker as we fight with CM over what's the current selection). Alt+Drag
does collide with rectangular-selection dragging, but that's less common &
in many cases it's easy to start the drag 1 char away from a number instead.

- Change nudge to Ctrl-Alt-arrow since Shift-Alt-arrow is now taken by "Add
Cursor to Next/Prev Line"

- Pass origin for selection changes as well to fix batching, since CM now
tracks selection changes interleaved in the undo history.